+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Add `:ruby_vips` analyzer to `store_dimensions` plugin (@janko-m)
+
 * Add `:mini_magick` analyzer to `store_dimensions` plugin (@janko-m)
 
 * Soft-rename `:heroku` logging format to `:logfmt` (@janko-m)